Sensors was merged into drive train
