+++ src/main/java/org/usfirst/frc/team1758/robot/commands/ApproachPeg.java
@@ -14,17 +14,16 @@ public class ApproachPeg extends CommandBase {
     logger.debug("ApproachPeg command created");
     requires(vision);
     requires(driveTrain);
-    requires(sensors);
   }
 
   protected void initialize() {
-    sensors.resetGyroAngle();
+    driveTrain.resetGyroAngle();
     finished = false;
     counter = 0;
   }
 
   protected void execute() {
-    logger.trace("Gyro: {} Center: {}", sensors.getUltrasonicValue(), isCentered());
+    logger.trace("Gyro: {} Center: {}", driveTrain.getUltrasonicValue(), isCentered());
     if (isDone()) {
       finished = true;
       logger.trace("ApproachPeg done");
@@ -37,7 +36,7 @@ public class ApproachPeg extends CommandBase {
 
   public void iterate() {
     double y = 0; 
-    if (sensors.getUltrasonicValue() > 40) {
+    if (driveTrain.getUltrasonicValue() > 40) {
       y = 0.3;
     }
     double x = 0;
@@ -45,7 +44,7 @@ public class ApproachPeg extends CommandBase {
       x = (vision.getCenterX() - RobotMap.CAMERA_WIDTH / 2) / (-2 * (RobotMap.CAMERA_WIDTH / 2));
     }
     double rotate = 0;
-    logger.trace("Angle: {}", .3 * sensors.getGyroAngle());
+    logger.trace("Angle: {}", .3 * driveTrain.getGyroAngle());
     driveTrain.mecanumDriveCartesian(.75 * x,.75 * y, rotate, 0);
   }
 
@@ -62,7 +61,7 @@ public class ApproachPeg extends CommandBase {
   }
 
   private boolean isDone() {
-    if (sensors.getUltrasonicValue() < 40) {
+    if (driveTrain.getUltrasonicValue() < 40) {
       counter++;
     } else {
       counter = 0;
