+++ src/main/java/org/usfirst/frc/team1758/robot/commands/MoveForward.java
@@ -16,7 +16,7 @@ public class MoveForward extends CommandBase {
   protected void initialize() {
     finished = false;
     driveTrain.resetEncoderPosition();
-    sensors.resetGyroAngle();
+    driveTrain.resetGyroAngle();
   }
 
   protected void execute() {
