+++ src/main/java/org/usfirst/frc/team1758/robot/commands/TurnLeft.java
@@ -5,7 +5,6 @@ public class TurnLeft extends CommandBase {
   private double angle;
 
   public TurnLeft(double angle) {
-    requires(sensors);
     requires(driveTrain);
     this.angle = angle;
   }
@@ -13,11 +12,11 @@ public class TurnLeft extends CommandBase {
   protected void initialize() {
     finished = false;
     driveTrain.resetEncoderPosition();
-    sensors.resetGyroAngle();
+    driveTrain.resetGyroAngle();
   }
 
   protected void execute() {
-    if (sensors.getGyroAngle() < -1.0 * this.angle) {
+    if (driveTrain.getGyroAngle() < -1.0 * this.angle) {
       finished = true;
       driveTrain.tankDrive(0, 0);
     } else {
